moved ddl, mocked batch func, updated createLogs calls after provisioning caching
