+++ runner/src/indexer-logger/indexer-logger.test.ts
@@ -37,8 +37,6 @@ describe('IndexerLogger', () => {
 
   describe('writeLog', () => {
         it('should call the query method with the correct parameters', async () => {
-            
-            const spyFormatDate = jest.spyOn(IndexerLogger.prototype, 'formatDate');
 
             const indexerLogger = IndexerLogger.create(mockDatabaseConnectionParameters, pgClient);
             const blockHeight = 123;
@@ -47,6 +45,8 @@ describe('IndexerLogger', () => {
             const logType = 'system';
             const logLevel = LogLevel.INFO;
             const message = 'Test message';
+            
+            const logDate = new Date(logTimestamp.getFullYear(), logTimestamp.getMonth(), logTimestamp.getDate());
 
             await indexerLogger.writeLog(
                 blockHeight,
@@ -57,12 +57,11 @@ describe('IndexerLogger', () => {
                 message
             );
 
-            expect(spyFormatDate).toHaveBeenCalledWith(logTimestamp);
             expect(query.mock.calls.length).toBe(1);
             expect(query.mock.calls[0][0]).toEqual('INSERT INTO testFunction.__logs (block_height, log_date, log_timestamp, log_type, log_level, message) VALUES ($1, $2, $3, $4, $5, $6)');
             expect(query.mock.calls[0][1]).toEqual([
                 blockHeight,
-                expect.any(String),  //toDo check if func name is correct
+                logDate,
                 logTimestamp,
                 logType,
                 LogLevel[logLevel],
